requirements.txt
langchain-cli
openai
langchain-community
langchain-core
streamlit
python-dotenv
chromaDB
PyMuPDF